% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{comma_cl}
\alias{comma_cl}
\alias{percent_cl}
\title{Local version of scales::comma scales::percent}
\usage{
comma_cl(
  x,
  accuracy = NULL,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = ".",
  decimal.mark = ",",
  trim = TRUE,
  digits,
  ...
)

percent_cl(
  x,
  accuracy = NULL,
  scale = 100,
  prefix = "",
  suffix = "\%",
  big.mark = ".",
  decimal.mark = ",",
  trim = TRUE,
  digits,
  ...
)
}
\arguments{
\item{x}{A numeric vector to format.}

\item{accuracy}{NULL}

\item{scale}{1}

\item{prefix}{""}

\item{suffix}{""}

\item{big.mark}{"."}

\item{decimal.mark}{","}

\item{trim}{TRUE}

\item{digits}{digits}

\item{...}{Other arguments passed on to base::format().}
}
\description{
Local version of scales::comma scales::percent
}
\examples{

x <- 10^c(1:10)

comma_cl(x)

comma_cl(x + .123, accuracy = 0.001)

comma_cl(x + .123, accuracy = 0.001, big.mark = " ")

scales::comma(x)

}
