% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcats.R
\name{fct_to_excel_values}
\alias{fct_to_excel_values}
\title{A function to apply before export values to excel}
\usage{
fct_to_excel_values(x)
}
\arguments{
\item{x}{A factor}
}
\description{
A function to apply before export values to excel
}
\examples{

f <- factor(c("b", "b", "a", "c", "c", "c"))

fct_to_excel_values(f)



}
