% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filename_gen}
\alias{filename_gen}
\title{Auxiliar function to generate a filename following a pattern}
\usage{
filename_gen(
  folder = "plots",
  pattern = "plot",
  ext = "png",
  npad = 3,
  verbose = TRUE
)
}
\arguments{
\item{folder}{.}

\item{pattern}{.}

\item{ext}{.}

\item{npad}{.}

\item{verbose}{npad .}
}
\description{
Auxiliar function to generate a filename following a pattern
}
