% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_geom_boxplot2.R
\name{geom_boxplot2}
\alias{geom_boxplot2}
\title{A box and whiskers plot (in the style of Tukey)}
\usage{
geom_boxplot2(
  mapping = NULL,
  data = NULL,
  stat = "boxplot",
  position = "dodge2",
  ...,
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  show.errorbar = TRUE,
  width.errorbar = 0.7,
  notch = FALSE,
  notchwidth = 0.5,
  varwidth = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{...}{...}

\item{outlier.colour}{outlier.colour}

\item{outlier.color}{outlier.color}

\item{outlier.fill}{outlier.fill}

\item{outlier.shape}{outlier.shape}

\item{outlier.size}{outlier.size}

\item{outlier.stroke}{outlier.stroke}

\item{outlier.alpha}{outlier.alpha}

\item{show.errorbar}{boolean}

\item{width.errorbar}{width of errorbar.}

\item{notch}{If `FALSE` (default) make a standard box plot. If
`TRUE`, make a notched box plot. Notches are used to compare groups;
if the notches of two boxes do not overlap, this suggests that the medians
are significantly different.}

\item{notchwidth}{For a notched box plot, width of the notch relative to
the body (defaults to `notchwidth = 0.5`).}

\item{varwidth}{If `FALSE` (default) make a standard box plot. If
`TRUE`, boxes are drawn with widths proportional to the
square-roots of the number of observations in the groups (possibly
weighted, using the `weight` aesthetic).}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}
}
\description{
All credits to @kongdd https://github.com/rpkgs/Ipaper
}
\examples{

df <- data.frame(
  x = sample(LETTERS[1:3], prob = 1:3, size = 1000, replace = TRUE),
  y = rgamma(1000, shape = 1)
)

df <- data.frame(
  x = sample(LETTERS[1:3], prob = 1:3, size = 1000, replace = TRUE),
  y = rt(1000, df = 3)
)

ggplot(df) +
  geom_boxplot(aes(x, y))

ggplot(df) +
  geom_boxplot2(aes(x, y))

ggplot(df) +
  geom_boxplot2(aes(x, y), width.errorbar = 0.2, width = 0.25)

}
