% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_point2}
\alias{geom_point2}
\alias{geom_jitter2}
\title{Geom point/jitter with shape 21 and color white}
\usage{
geom_point2(..., shape = 21, color = "white")

geom_jitter2(..., shape = 21, color = "white")
}
\arguments{
\item{...}{geom_point/geom_jitter parameters.}

\item{shape}{Shape 21}

\item{color}{Color "white"}
}
\description{
Geom point/jitter with shape 21 and color white
}
\examples{

ggplot(iris) +
  geom_jitter(aes(Sepal.Length, Sepal.Width, color = Species))

ggplot(iris) +
  geom_jitter2(aes(Sepal.Length, Sepal.Width, fill = Species))

ggplot(iris) +
  geom_point(aes(Sepal.Length, Sepal.Width, color = Species))

ggplot(iris) +
  geom_point2(aes(Sepal.Length, Sepal.Width, fill = Species))

}
