% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_geom_boxplot2.R
\docType{data}
\name{GeomBoxplot2}
\alias{GeomBoxplot2}
\title{No description}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{...}{...}

\item{outlier.colour}{outlier.colour}

\item{outlier.color}{outlier.color}

\item{outlier.fill}{outlier.fill}

\item{outlier.shape}{outlier.shape}

\item{outlier.size}{outlier.size}

\item{outlier.stroke}{outlier.stroke}

\item{outlier.alpha}{outlier.alpha}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}
}
\description{
No description
}
\keyword{datasets}
