% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{jbkverse_deps}
\alias{jbkverse_deps}
\title{List all jbkverse dependencies}
\usage{
jbkverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
jbkverse packages.}
}
\description{
List all jbkverse dependencies
}
