% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{jbkverse_packages}
\alias{jbkverse_packages}
\title{List all packages in the jbkverse}
\usage{
jbkverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include jbkverse in the list?}
}
\description{
List all packages in the jbkverse
}
\examples{
jbkverse_packages()
}
