% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{jbkverse_update}
\alias{jbkverse_update}
\title{Update jbkverse packages}
\usage{
jbkverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
jbkverse packages.}
}
\description{
This will check to see if all jbkverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
jbkverse_update()
}
}
