% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{kunstomverse_packages}
\alias{kunstomverse_packages}
\title{List all packages in the kunstomverse}
\usage{
kunstomverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include kunstomverse in the list?}
}
\description{
List all packages in the kunstomverse
}
\examples{
kunstomverse_packages()
}
