% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{kunstomverse_update}
\alias{kunstomverse_update}
\title{Update kunstomverse packages}
\usage{
kunstomverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
kunstomverse packages.}
}
\description{
This will check to see if all kunstomverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
kunstomverse_update()
}
}
