% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_sleep}
\alias{progress_sleep}
\title{Wrapper of `Sys.sleep()` using  `{progress}` package}
\usage{
progress_sleep(time, format = "  waiting [:bar] :percent eta: :eta", ...)
}
\arguments{
\item{time}{The time interval to suspend execution for, in seconds.}

\item{format}{The format of the progress bar. A number of tokens can be 
used here, see them below. It defaults to "[:bar] :percent", which means 
that the progress bar is within brackets on the left, and the percentage 
is printed on the right. Source: progress package.}

\item{...}{Parameters for \code{progress::progress_bar$new}}
}
\description{
Wrapper of `Sys.sleep()` using  `{progress}` package
}
\examples{

progress_sleep(5)

system.time(progress_sleep(10))

}
