% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiktok.R
\name{tik}
\alias{tik}
\alias{tok}
\alias{tiktok_clear}
\alias{tik_msg}
\alias{tok_msg}
\alias{tiktok_log}
\title{tic toc Wrappers}
\usage{
tik(msg = NULL, quiet = FALSE, func.tik = tik_msg)

tok(log = TRUE, func.tok = tok_msg)

tiktok_clear()

tik_msg(tik, msg = NULL)

tok_msg(tik, tok, msg, info)

tiktok_log()
}
\arguments{
\item{msg}{msg}

\item{quiet}{quiet}

\item{func.tik}{func.tik}

\item{log}{log}

\item{func.tok}{func.tok}

\item{tik}{tik}

\item{tok}{tok}

\item{info}{info}
}
\description{
tic toc Wrappers
}
\examples{

tiktok_clear()

tik(msg = "Go to sleep")

Sys.sleep(2)

tok()

tik(msg = "Go to sleep again!")

Sys.sleep(1)

tok()

tiktok_log()

}
