% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_knst.R
\name{theme_knst}
\alias{theme_knst}
\title{My ggplot2 go-to theme inspired by \code{hrbrthemes}}
\usage{
theme_knst(...)
}
\arguments{
\item{...}{Parameters for \code{hrbrthemes::theme_ipsum}}
}
\description{
My ggplot2 go-to theme inspired by \code{hrbrthemes}
}
\examples{

\dontrun{

library(ggplot2)
 library(extrafont)
 library(hrbrthemes)

p <- ggplot(mtcars) +
   geom_point(aes(mpg, hp, color = factor(carb)), size = 2.5) +
   hrbrthemes::scale_color_ipsum() +
   labs(x="Fuel effiiency (mpg)", y="Weight (tons)",
        title="Seminal ggplot2 scatterplot example",
        subtitle="A plot that is only useful for demonstration purposes",
        caption="Brought to you by the letter 'g'") +
  theme_knst()
  
p
 
 p + 
   facet_wrap(vars(factor(carb)))
} 

}
